feat(sdk): updates go sdk paths and build configuration
