+++ sdk/go/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/config"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/config"
 )
 
 func main() {
