+++ sdk/go/pkg/artifact/gobin.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/ALT-F4-LLC/vorpal/sdk/go/api/v0/artifact"
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/config"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/config"
 )
 
 func GoBin(context *config.ConfigContext) (*string, error) {
