+++ sdk/go/pkg/artifact/language/rust.go
@@ -10,8 +10,8 @@ import (
 	"text/template"
 
 	artifactApi "github.com/ALT-F4-LLC/vorpal/sdk/go/api/v0/artifact"
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/artifact"
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/config"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/artifact"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/config"
 	"github.com/BurntSushi/toml"
 )
