+++ sdk/go/pkg/artifact/protoc_gen_go_grpc.go
@@ -4,23 +4,23 @@ import (
 	"errors"
 
 	"github.com/ALT-F4-LLC/vorpal/sdk/go/api/v0/artifact"
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/config"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/config"
 )
 
-func Protoc(context *config.ConfigContext) (*string, error) {
+func ProtocGenGoGRPC(context *config.ConfigContext) (*string, error) {
 	target := context.GetTarget()
 
 	var digest string
 
 	switch target {
 	case artifact.ArtifactSystem_AARCH64_DARWIN:
-		digest = "8ad451bdcda8f24f4af59ccca23fd71a06975a9d069571f19b9a0d503f8a65c8"
+		digest = "410549ec26b1b169c64ca0d4b6c09987000b0e88b7854a608708806c58a13dcc"
 	case artifact.ArtifactSystem_AARCH64_LINUX:
-		digest = "d0f3c08c81bebdb5a502863c786a03d661e4faad1941e941f705bb076eaff13c"
+		digest = "0197068bfea81502d1e152c4bc4c4e5584c191d5931b9f68dc1ac5f3aa9a67a4"
 	case artifact.ArtifactSystem_X8664_DARWIN:
-		digest = "2a3d7816b06f1a046cbf8e82c1a94fe71b4fd384726f2064c9e0960ac75dadec"
+		digest = "ee3fee174a60350ba21e971557b2ae189fc7674127d6e528fd78aa8d151d98c8"
 	case artifact.ArtifactSystem_X8664_LINUX:
-		digest = "c5bded4de6ca52ac5e731e328e22b5dfb957009ed5a553e24ba9fdec4379ba44"
+		digest = "ed50829e4f561a392881bb81b10c6617e7ecf08a0dc8a06e9515c559f51f3ffa"
 	default:
 		return nil, errors.New("unsupported target")
 	}
