+++ sdk/go/pkg/artifact/step.go
@@ -7,7 +7,7 @@ import (
 	"text/template"
 
 	artifactApi "github.com/ALT-F4-LLC/vorpal/sdk/go/api/v0/artifact"
-	"github.com/ALT-F4-LLC/vorpal/sdk/go/internal/config"
+	"github.com/ALT-F4-LLC/vorpal/sdk/go/pkg/config"
 )
 
 type BashScriptTemplateArgs struct {
