+++ sdk/go/pkg/store/hash.go
